--- README.md
@@ -1,2 +1,0 @@
-# Android-CIM
- 输入法快捷更换
