--- app/src/main/res/drawable/corner.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle" >
-    <padding
-        android:bottom="5dp"
-        android:left="5dp"
-        android:right="5dp"
-        android:top="5dp"
-        />
-    <!-- 设置圆角矩形 -->
-    <corners android:radius="60dp" />
-    <stroke
-        android:width="1dp"
-        android:color="#FFFFFF" />
-    <solid android:color="#FFFFFF" />
-
-</shape>
