--- app/src/main/res/drawable/image_button_background.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item  android:state_pressed="false">
-        <shape
-            android:shape="rectangle" >
-            <padding
-                android:bottom="5dp"
-                android:left="5dp"
-                android:right="5dp"
-                android:top="5dp"
-                />
-            <!-- 设置圆角矩形 -->
-            <corners android:radius="30dp" />
-            <stroke
-                android:width="1dp"
-                android:color="@color/white" />
-            <solid android:color="@color/white" />
-
-        </shape>
-    </item>
-    <item  android:state_pressed="true">
-        <shape
-            android:shape="rectangle" >
-            <padding
-                android:bottom="5dp"
-                android:left="5dp"
-                android:right="5dp"
-                android:top="5dp"
-                />
-            <!-- 设置圆角矩形 -->
-            <corners android:radius="30dp" />
-            <stroke
-                android:width="1dp"
-                android:color="@color/blue3" />
-            <solid android:color="@color/blue3" />
-
-        </shape>
-    </item>
-</selector>
