--- app/src/main/res/drawable/image_button_foreground.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ripple xmlns:android="http://schemas.android.com/apk/res/android"
-    android:color="#22000000">
-    <item android:id="@android:id/mask">
-
-        <shape android:shape="rectangle">
-
-            <corners android:radius="30dp" />
-
-            <solid android:color="?android:attr/colorPrimary" />
-        </shape>
-    </item>
-</ripple>
