--- app/src/main/res/drawable/shade.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle"
-    >
-
-    <gradient
-        android:angle="90"
-        android:gradientRadius="10dp"
-        android:endColor="@color/blue2"
-        android:startColor="@color/blue3"
-        android:type="linear"
-        android:useLevel="true" />
-
-<!--    <solid android:color="#FFFFFF" >-->
-
-<!--    </solid>-->
-</shape>
