+++ app/src/main/res/layout/activity_main.xml
@@ -1,89 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    tools:context=".MainActivity">
 
-    android:orientation="vertical"
-    android:background="@color/blue3">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="8dp"
-        android:gravity="center"
-        android:background="@color/blue2">
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Hello World!"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-    </LinearLayout>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="3"
-        android:orientation="vertical"
-        android:layout_margin="30dp"
-        android:background="@drawable/corner"
-        android:gravity="center"
-        >
-        <Button
-            android:id="@+id/button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-
-            android:text="切换输入法"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/textView"
-            app:layout_constraintVertical_bias="0.3" />
-
-
-        <ImageButton
-            android:id="@+id/image_button"
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            android:background="@drawable/image_button_background"
-            android:scaleType="centerInside"
-            android:layout_margin="10dp"
-
-            android:foreground="@drawable/image_button_foreground"
-            android:src="@drawable/ic_keyboard"
-
-            />
-
-        <Button
-            android:id="@+id/button2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="设置Tile"
-
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/button"
-            app:layout_constraintVertical_bias="0.17000002" />
-<!--        <EditText-->
-<!--            android:layout_width="match_parent"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:text="sdlfkj"/>-->
-
-    </LinearLayout>
-    <LinearLayout
-        android:id="@+id/l_touch"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:gravity="center">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="\@bilibili\n 易名亦名"
-            android:textColor="@color/blue1"
-            android:textSize="20sp"
-            android:textAlignment="center"
-            >
-        </TextView>
-    </LinearLayout>
-
-
-
-
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
