+++ app/src/main/res/values-night/themes.xml
@@ -1,12 +1,7 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.ACIM" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.ACIMApplication" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your dark theme here. -->
         <!-- <item name="colorPrimary">@color/my_dark_primary</item> -->
-    </style>
-
-    <style name="ThemeOverlay.ACIM.FullscreenContainer" parent="">
-        <item name="fullscreenBackgroundColor">@color/light_blue_900</item>
-        <item name="fullscreenTextColor">@color/light_blue_A400</item>
     </style>
 </resources>
