--- app/src/main/res/values/attrs.xml
@@ -1,6 +1,0 @@
-<resources>
-    <declare-styleable name="FullscreenAttrs">
-        <attr name="fullscreenBackgroundColor" format="color" />
-        <attr name="fullscreenTextColor" format="color" />
-    </declare-styleable>
-</resources>
