+++ app/src/main/res/values/themes.xml
@@ -1,20 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.ACIM" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.ACIMApplication" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.ACIM" parent="Base.Theme.ACIM" />
-
-    <style name="Theme.ACIM.Fullscreen" parent="Theme.ACIM">
-        <item name="android:actionBarStyle">@style/Widget.Theme.ACIM.ActionBar.Fullscreen</item>
-        <item name="android:windowActionBarOverlay">true</item>
-        <item name="android:windowBackground">@null</item>
-    </style>
-
-    <style name="ThemeOverlay.ACIM.FullscreenContainer" parent="">
-        <item name="fullscreenBackgroundColor">@color/light_blue_600</item>
-        <item name="fullscreenTextColor">@color/light_blue_A200</item>
-    </style>
+    <style name="Theme.ACIMApplication" parent="Base.Theme.ACIMApplication" />
 </resources>
